requirements.txt
anyio             4.4.0
astor             0.8.1
asttokens         2.4.1
backcall          0.2.0
certifi           2024.7.4
coloredlogs       15.0.1
decorator         5.1.1
exceptiongroup    1.2.2
executing         2.0.1
flatbuffers       24.3.25
h11               0.14.0
httpcore          1.0.5
httpx             0.27.0
humanfriendly     10.0
idna              3.7
ipdb              0.13.13
ipython           8.12.3
jedi              0.19.1
markdown-it-py    3.0.0
matplotlib-inline 0.1.7
mdurl             0.1.2
mpmath            1.3.0
numpy             1.24.4
onnx              1.16.1
onnx-simplifier   0.4.36
onnx-tool         0.9.0
onnxruntime       1.18.1
opencv-python     3.4.8.29
opt-einsum        3.3.0
packaging         24.1
paddle2onnx       1.2.6
paddlepaddle-gpu  2.6.1
parso             0.8.4
pexpect           4.9.0
pickleshare       0.7.5
pillow            10.4.0
pip               24.0
prompt_toolkit    3.0.47
protobuf          5.27.2
ptyprocess        0.7.0
pure-eval         0.2.2
pyclipper         1.3.0.post5
Pygments          2.18.0
rich              13.7.1
setuptools        69.5.1
shapely           2.0.5
six               1.16.0
sniffio           1.3.1
stack-data        0.6.3
sympy             1.13.0
tabulate          0.9.0
tomli             2.0.1
traitlets         5.14.3
typing_extensions 4.12.2
wcwidth           0.2.13
wheel             0.43.0